+++ lib/providers/place_provider.dart
@@ -8,7 +8,6 @@ import 'package:google_maps_place_picker_mb/src/place_picker.dart';
 import 'package:google_maps_webservice/geocoding.dart';
 import 'package:google_maps_webservice/places.dart';
 import 'package:http/http.dart';
-import 'package:location/location.dart' as LocationPlatformInterface;
 import 'package:provider/provider.dart';
 import 'package:get/get.dart';
 
@@ -47,6 +46,7 @@ class PlaceProvider extends ChangeNotifier {
   bool isAutoCompleteSearching = false;
   CircleArea? searchRadiusArea;
 
+<<<<<<< HEAD
   LocationPlatformInterface.Location location =
       new LocationPlatformInterface.Location();
   LocationPlatformInterface.PermissionStatus permissionGranted =
@@ -107,6 +107,43 @@ class PlaceProvider extends ChangeNotifier {
       print(e);
       currentPosition = null;
     }
+=======
+  Future<void> updateCurrentLocation() async {
+    bool serviceEnabled;
+    LocationPermission permission;
+
+    // Test if location services are enabled.
+    serviceEnabled = await Geolocator.isLocationServiceEnabled();
+    if (!serviceEnabled) {
+      // Location services are not enabled don't continue
+      // accessing the position and request users of the
+      // App to enable the location services.
+      return Future.error('Location services are disabled.');
+    }
+
+    permission = await Geolocator.checkPermission();
+    if (permission == LocationPermission.denied) {
+      permission = await Geolocator.requestPermission();
+      if (permission == LocationPermission.denied) {
+        // Permissions are denied, next time you could try
+        // requesting permissions again (this is also where
+        // Android's shouldShowRequestPermissionRationale
+        // returned true. According to Android guidelines
+        // your App should show an explanatory UI now.
+        return Future.error('Location permissions are denied');
+      }
+    }
+
+    if (permission == LocationPermission.deniedForever) {
+      // Permissions are denied forever, handle appropriately.
+      return Future.error(
+          'Location permissions are permanently denied, we cannot request permissions.');
+    }
+
+    _currentPosition = await Geolocator.getCurrentPosition(
+      desiredAccuracy: desiredAccuracy ?? LocationAccuracy.best,
+    );
+>>>>>>> d92469e56bd389d0eda20bc28b303b4abfbd27e0
   }
 
   Position? _currentPosition;